Fixes trouble with fast movement bug, adds comments, cleans up
Fast movement bug: The staggering of row movement was happening
more slowly than the movement for all the rows was called (when
there were only a few ai left). This caused ai to move beyond
the bounds of the .game-container, freezing ai over one side.
TO SUM: the staggering of the movement of rows relative to each
other must always be completed before the next movement of all
rows is called (so speed up the staggering based on all movement).